backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.12",
    "yfinance>=0.2.55",
    "sqlmodel>=0.0.14",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
    "cryptography>=42.0.0",
    "pyjwt>=2.10.1",
    "bcrypt==4.0.1",
    "elasticsearch==8.*",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "pytest>=8.3.5",
    "types-passlib>=1.7.7.20250408",
]


[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 140
